requirements_stable.txt
# Stable Requirements for PetBot - All Versions Tested Together
# This ensures compatibility and avoids dependency hell

# Core ML and NLP
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==5.1.1
transformers==4.57.1
torch==2.1.2
huggingface-hub==0.19.4

# Text processing
nltk==3.8.1
rapidfuzz==3.6.1
joblib==1.3.2

# RAG and Vector Search
chromadb==0.4.22
rank-bm25==0.2.2
faiss-cpu==1.7.4

# LLM Integration
requests==2.31.0
openai==1.12.0

# Web Framework
streamlit==1.28.1

# Data Processing (lightweight alternatives)
# Note: pandas requires pyarrow which needs cmake
# We'll use built-in Python data structures instead
# pandas==2.0.3  # Commented out due to cmake requirement

# Azure Integration
azure-storage-blob==12.19.1
azure-identity==1.15.0
azure-core==1.29.5

# Advanced ML (for Azure components)
faiss-cpu==1.11.0
rapidfuzz==3.6.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
